transition.py: move to_utc_string() to common.py
